Remove null toolchain
Close #9

Move nulltoolchain to test scope. Maybe we'll need it in the future.
